Add polling-based project log code (#204)
Fixes #188 (in an MVP form)

This PR covers the REST version of fetching docker logs, which is the easy
one that comes with built in "starting at what datetime?" functionality.
Hooking websockets into the docker log ouput so it updates if and when
there's more log data is going to need more work, and should not hold
up landing this, because really any sort of logs insight is better than no
logs insight.